requirements.txt
fastapi
uvicorn
twilio
openai
pyttsx3
python-dotenv
pydantic
